Struct terminal_trx::Terminal

pub struct Terminal(/* private fields */);

A readable and writable handle to the terminal (or TTY), created using terminal(). You can read and write data using the io::Read and io::Write implementations respectively.

Use Terminal::lock if you want to avoid locking before each read / write call.

Implementations

impl Terminal

pub fn lock(&mut self) -> TerminalLock<'_>

Locks access to this terminal, returing a guard that is readable and writable.

Until the returned TerminalLock is dropped, all standard I/O streams that refer to the same terminal will be locked.